requirements.txt
dnspython
Flask
Flask-Bcrypt
Flask-Login
matplotlib
mysql-connector-python
PyMySQL
pyOpenSSL
python-dotenv
PyYAML
requests
